testttt

+++ SpringEltpCaseStudy/src/main/java/com/spring/springcore/dao/EmployeeDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.spring.springcore.entity.Employee;
 
-public interface EmployeeDAO extends JpaRepository<Employee, Integer>{
+public interface EmployeeDAO extends JpaRepository<Employee, Long>{
 
 }
